+++ src/anycase_fork/cobol.nim
@@ -1,0 +1,6 @@
+from strutils import toUpperAscii
+
+import kebab
+
+func cobol*(str: string): string =
+  result = kebab(str).toUpperAscii()