+++ src/anycase_fork/kebab.nim
@@ -1,0 +1,6 @@
+from strutils import join
+
+import words
+
+func kebab*(str: string): string =
+  result = join(words(str), "-")