+++ src/anycase_fork/pascal.nim
@@ -1,9 +1,10 @@
-import anycase/words
 from strutils import join, capitalizeAscii
 from sequtils import map
 
-proc pascal*(str: string): string =
+import words
+
+func pascal*(str: string): string =
   let parts = words(str)
   let capitalizedParts = map(parts, capitalizeAscii)
 
-  return join(capitalizedParts)
+  result = join(capitalizedParts)