+++ src/anycase_fork/path.nim
@@ -1,0 +1,8 @@
+from strutils import join
+
+import words
+
+func path*(str: string): string =
+  let parts = words(str)
+
+  result = join(parts, "/")