+++ src/anycase_fork/plain.nim
@@ -1,0 +1,8 @@
+from strutils import join
+
+import words
+
+func plain*(str: string): string =
+  let parts = words(str)
+
+  result = join(parts, " ")