+++ src/anycase_fork/train.nim
@@ -1,0 +1,11 @@
+from strutils import join, capitalizeAscii
+from sequtils import map
+
+import words
+
+func train*(str: string): string =
+  let parts = words(str)
+  let capitalizedParts = map(parts, capitalizeAscii)
+
+  result = join(capitalizedParts, "-")
+