+++ src/anycase_fork/upper_snake.nim
@@ -1,0 +1,6 @@
+from strutils import toUpperAscii
+
+import snake
+
+func upperSnake*(str: string): string =
+  result = snake(str).toUpperAscii()