+++ tests/test_compile_time.nim
@@ -1,0 +1,23 @@
+discard """
+  exitcode: 0
+"""
+
+import anycase_fork
+
+discard static(camel("test"))
+
+discard static(cobol("test"))
+
+discard static(kebab("test"))
+
+discard static(pascal("test"))
+
+discard static(path("test"))
+
+discard static(plain("test"))
+
+discard static(snake("test"))
+
+discard static(train("test"))
+
+discard static(upperSnake("test"))