+++ tests/test_kebab.nim
@@ -1,22 +1,23 @@
-import unittest
-import anycase/kebab
+discard """
+  exitcode: 0
+"""
 
-suite "kebab":
-  test "plain -> kebab":
-    check kebab("change my case") == "change-my-case"
+import anycase_fork
 
-  test "path -> kebab":
-    check kebab("change/my/case") == "change-my-case"
+assert kebab("change my case") == "change-my-case"
 
-  test "kebab -> kebab":
-    check kebab("change-my-case") == "change-my-case"
+assert kebab("change/my/case") == "change-my-case"
 
-  test "camel -> kebab":
-    check kebab("changeMyCase") == "change-my-case"
+assert kebab("change-my-case") == "change-my-case"
 
-  test "pascal -> kebab":
-    check kebab("ChangeMyCase") == "change-my-case"
+assert kebab("changeMyCase") == "change-my-case"
 
-  test "snake -> kebab":
-    check kebab("change_my_case") == "change-my-case"
+assert kebab("ChangeMyCase") == "change-my-case"
 
+assert kebab("change_my_case") == "change-my-case"
+
+assert kebab("CHANGE-MY-CASE") == "change-my-case"
+
+assert kebab("Change-My-Case") == "change-my-case"
+
+assert kebab("CHANGE_MY_CASE") == "change-my-case"