+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v0.2.0
+
+- Use regex instead of std for compile-time
+- New cases:
+    - `cobol`
+    - `train`
+    - `upperSnake`
+- Use testament instead of unittest
+
 # v0.1.0
 
 Base library implementation.